Remove `404.html`
Providing a 404.html page doesn't make much sense for this project
as users will need to configure their server in order for the page
to be actually used.

In HTML5 Boilerplate, the idea is that the server configs (Apache or
the alternatives) will basically ensure that the user's server will
return the `404.html` page in case of an 404 error.

Also, with HTML5 Boilerplate, we've seen that users quite often forget
to change/replace the default 404 page.