Fixes #337 - restore touch icons on Android && iOS
After extensive testing in the office, we’ve discovered possible bugs
in Chrome’s icon cache to do with both the `sizes` property and how
well we handle icons of size `196x196`. We tested icons at `195x195` in
case these passed, but sadly this new set of configuration was the only
one we could reliably get working.

Long-term, we’ll want to move to using the recommendations in the
official Chrome for Android docs around add to homescreen, but at
present they don’t represent a setup that works. This setup does work
giving you icons on Android and iOS as you would expect.