Updated ComponentHandler for many class registrations
Updated ComponentHandler to handle multiple class registrations for the
same element. Now, instead of just having data-upgraded attribute set,
it adds the classname you are upgrading to, to this attribute. Thus for
things like the button demo which is an instance of both button AND
ripple - it can be upgraded twice with 2 different classes handling the
same element. Woohoo!