fix(fab): Allow any child of a fab to align properly in all browsers.
Some browsers don't want to align an SVG child in the center. This patch allows for anything given as a child to have full width which makes all engines align them to the center of the fab. Universal selector here also allows for less overhead needing to add in an extra class for inner nodes.