fix(scripts): Update determine-pkg-versions to use new pkg names (#141)
determine-pkg-versions was erroring because it was assumes the `name`
key in each component's package.json to be `mdc-*`, rather than
`@material/*`. This commit fixes that issue.

Additionally, it allows _all_ commit types to constitute at least a
patch version bump. For example, a docs change, chore, or non-breaking
refactor to a component could all be considered worthy of a patch
version update. This ensures that our latest and greatest code and docs
are available on npm whenever we release.

[ci skip]