fix(demos): Sanitize slider input values (#2018)
The demo page now checks that the user's input values are finite numbers before assigning them to any of the component's setter properties. This prevents an error from being thrown when the user deletes all the text in one of the inputs (which is a fairly common thing to do before entering a new number).

In addition, to make it clearer that the slider supports negative numbers for `min` and `max`, the demo page now allows those input values to go down to `-100` (previously they had a lower bound of `0`).