chore(infrastructure): Update screenshots after browser upgrades (#4406)
Firefox updated from version 63 to 65, which slightly changed antialiasing behavior of text
and `border-radius`.

Unfortunately, Chrome 72 broke `padding-right` on `mdc-dialog__actions` for `mdc-dialog--stacked`:

https://storage.googleapis.com/mdc-web-screenshot-tests/advorak/2019/02/11/22_13_08_557/report/report.html

This may be related to https://crbug.com/927066. It still renders correctly in Firefox 65, IE 11,
and Edge 18.

I wasn't able to find a workaround, so I'm approving the broken "stacked dialog" screenshots in Chrome for now. If/when Chromium fixes this regression, the dialog screenshot tests will start failing, which will remind us to update the screenshots again.