fix(floatinglabel): Estimate hidden scroll width (#5448)
In some cases, the floating label needs to immediately know its width. This creates problems if the floating label is instantiated inside a display: none; parent element, like a hidden dialog. To resolve that, we provide a helper method that estimates the width of an element if hidden. If visible, it computes the true width.